Zonefile#resource_records: add PowerDNS-SQL export
The PowerDNS authors decided in 1999, they'd store data with
"no terminating dot in SQL storage". This became an "official
wart" [1], as it makes processing PDNS SQL data harder with
tools which expect well-formed RR data...

We need to interact in [2] with that storage and decided to
implement the data transformation closer to the raw data.

[1] https://github.com/PowerDNS/pdns/blob/rel/auth-4.2.x/pdns/zone2sql.cc#L59-L63
[2] https://github.com/digineo/dnsgit